requirements.txt
Pillow==9.5.0
streamlit==1.24.0
streamlit_option_menu==0.3.6
